% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_ios.R
\name{susie_ios}
\alias{susie_ios}
\title{Fine-Mapping and Estimation of Causal Variants Using SuSiE}
\usage{
susie_ios(
  z,
  R,
  n,
  L = 10,
  pip.threshold = 0.5,
  max.iter = 10,
  max.eps = 0.001,
  inner.iter = 5,
  score.test = F,
  estimate_residual_variance = F
)
}
\arguments{
\item{z}{Z scores of GWAS effect sizes.}

\item{R}{LD (Linkage Disequilibrium) matrix of variants.}

\item{n}{Sample size of GWAS data.}

\item{L}{Number of single effects to consider, default is 10.}

\item{pip.threshold}{A threshold to determine which variants are considered causal, default is 0.5.}

\item{max.iter}{Maximum number of iterations for estimating the infinitesimal effect, default is 15.}

\item{max.eps}{The maximum epsilon for convergence in the iterative process, default is 0.001.}

\item{inner.iter}{Number of iterations for the inner loop of the estimation process, default is 5.}

\item{score.test}{Perform score test of variance component or not, default to F.}

\item{estimate_residual_variance}{Boolean flag to indicate whether to estimate the residual variance, default is FALSE.}
}
\value{
A list containing the following elements:
  - \code{eta}: Linear predictor, which is the sum of the causal effect (\code{beta}) and the infinitesimal effect (\code{alpha}). It represents the total genetic effect of variants on the trait, combining the direct effects of causal variants and the background genetic effects.
  - \code{beta}: Causal effect. These are the effect estimates for variants identified with significant effect sizes through the SuSiE method, representing the direct impact of these variants on the trait.
  - \code{alpha}: Infinitesimal effect. This represents the aggregate effect of all other genetic variants in the background, excluding the identified causal variants. It can be considered as the unexplained genetic variability in the model.
  - \code{var.inf}: Variance of the infinitesimal effect. Estimated via REML, this reflects the degree of variation in the infinitesimal effect across the genetic variation.
  - \code{pv}: P-value from the score test. This value is used to assess whether the variance of the infinitesimal effect is significantly non-zero, serving as an important metric to test if the model adequately explains the genetic variability.
  - \code{pip}: Posterior Inclusion Probabilities for the estimates of causal effects. These are the probabilities calculated by the SuSiE method for each variant, assessing the likelihood that a variant is causal.
  - \code{fit.susie}: Output from the SuSiE method. This is the complete result object from the SuSiE method after performing fine-mapping, containing detailed information about the selected model and estimated parameters, allowing for further analysis and interpretation of the results.
}
\description{
This function utilizes the SuSiE (Sum of Single Effects) method for fine-mapping to identify causal variants with significant effect sizes. It employs REML (Restricted Maximum Likelihood) for the estimation of infinitesimal effects and for determining the optimal variance of these effects. Additionally, the function utilizes a score test to assess whether the variance of the infinitesimal effect is zero. The function has been updated to include parameters for controlling the precision of the estimation process and the option to estimate residual variance.
}
\details{
The function starts by using the `susie_rss` function for fine-mapping based on the input Z scores, LD matrix, and sample size to identify causal variants. It then filters these variants based on the PIP threshold to calculate their effect sizes. For the estimation of infinitesimal effects, if the score test's p-value is below the specified threshold, an iterative process is used to estimate the optimal variance of these effects. If the p-value does not meet the threshold, the variance of the infinitesimal effect is set to zero. This approach aims to enhance the precision of fine-mapping by selecting causal variants based on their effect sizes and accurately estimating the underlying genetic architecture through REML. The inclusion of parameters for controlling the precision of the estimation process and the option to estimate residual variance allows for more flexibility and accuracy in the analysis.
}
\examples{
# Example usage:
# Assuming z, R, and n are defined:
result <- susie_ios(z, R, n, L, pip.threshold, max.iter, max.eps, inner.iter, estimate_residual_variance)
}
