% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_ios_onestep.R
\name{susie_ios_onestep}
\alias{susie_ios_onestep}
\title{Fine-Mapping and Estimation of Causal Variants and One-Step Infinitesimal Effect Estimation Using SuSiE}
\usage{
susie_ios_onestep(
  z,
  R,
  n,
  L = 10,
  pip.threshold = 0.25,
  iter = 5,
  inner.iter = 3,
  score.test = F,
  pleiotropy.keep = "ALL",
  estimate_residual_variance = F
)
}
\arguments{
\item{z}{Z scores of GWAS effect sizes.}

\item{R}{LD (Linkage Disequilibrium) matrix of variants.}

\item{n}{Sample size of GWAS data.}

\item{L}{Number of single effects to consider, default is 10.}

\item{pip.threshold}{A threshold to determine which variants are considered causal, default is 0.25.}

\item{iter}{Maximum number of iterations for the one-step estimation process, default is 15.}

\item{inner.iter}{Number of iterations for the inner loop of the one-step estimation process, default is 5.}

\item{score.test}{Perform score test of variance component or not, default to F.}

\item{pleiotropy.keep}{The indices of variants to input for fine-mapping, default to "ALL".}

\item{estimate_residual_variance}{Boolean flag to indicate whether to estimate the residual variance in the one-step process, default is FALSE.}
}
\value{
A list containing the following elements:
  - \code{eta}: Linear predictor, which is the sum of the causal effect (\code{beta}) and the infinitesimal effect (\code{alpha}). It represents the total genetic effect of variants on the trait, combining the direct effects of causal variants and the background genetic effects.
  - \code{beta}: Causal effect. These are the effect estimates for variants identified with significant effect sizes through the SuSiE method, representing the direct impact of these variants on the trait.
  - \code{alpha}: Infinitesimal effect. This represents the aggregate effect of all other genetic variants in the background, excluding the identified causal variants. It can be considered as the unexplained genetic variability in the model.
  - \code{var.inf}: Variance of the infinitesimal effect. Estimated via REML in a one-step process, this reflects the degree of variation in the infinitesimal effect across the genetic variation.
  - \code{pip}: Posterior Inclusion Probabilities for the estimates of causal effects. These are the probabilities calculated by the SuSiE method for each variant, assessing the likelihood that a variant is causal.
  - \code{fit.susie}: Output from the SuSiE method. This is the complete result object from the SuSiE method after performing fine-mapping and the one-step infinitesimal effect estimation, containing detailed information about the selected model and estimated parameters, allowing for further analysis and interpretation of the results.
}
\description{
This function extends the SuSiE (Sum of Single Effects) method for fine-mapping to not only identify causal variants with significant effect sizes but also to perform a one-step estimation of the infinitesimal effect based on the residuals of the fine-mapping results. It employs REML (Restricted Maximum Likelihood) for the estimation of infinitesimal effects and for determining the optimal variance of these effects, following the identification of causal variants. This one-step approach aims to streamline the process by directly estimating the infinitesimal effect after the causal variants have been identified, enhancing the efficiency and accuracy of genetic effect estimation.
}
\details{
The function begins with the `susie_rss` method for fine-mapping based on the input Z scores, LD matrix, and sample size to identify causal variants. After identifying these variants, it utilizes a one-step process to estimate the infinitesimal effect based on the residuals. This process involves adjusting the estimation of the infinitesimal effect to account for the identified causal variants, thereby refining the overall genetic effect estimation. The inclusion of a one-step estimation process for the infinitesimal effect aims to enhance the precision and efficiency of genetic architecture estimation by directly incorporating the fine-mapping results into the infinitesimal effect estimation.
}
\examples{
# Example usage:
# Assuming z, R, and n are defined:
result <- susie_ios_onestep(z, R, n, L, pip.threshold, iter, inner.iter, estimate_residual_variance)
}
